Move map-related files/functions to the new map app.
