+++ project/lib/templates/lib/index.html
@@ -58,7 +58,7 @@
 
 </div><!-- /intro-->
 
-{% include 'lib/map.html' %}
+{% include 'map/map.html' %}
 
 <div class="index-container-after-map">
   <div class="site-stats">
@@ -75,7 +75,7 @@
   </div>
 </div>
 
-{% include 'lib/map_scripts.html' with map_sources=map_sources %}
+{% include 'map/map_scripts.html' with map_sources=map_sources %}
 
 <script type="text/javascript">
     $(document).ready(function(){
