+++ project/lib/templatetags/common_tags.py
@@ -2,7 +2,6 @@
 
 import datetime
 import json
-from urllib.parse import urlencode
 
 from django import template
 from django.conf import settings
@@ -16,20 +15,6 @@ register = template.Library()
 @register.simple_tag
 def get_form_media(form):
     return dict(js=form.media._js, css=form.media._css)
-
-
-@register.simple_tag
-def google_maps_api_url(callback):
-    url = 'https://maps.googleapis.com/maps/api/js'
-
-    url_kwargs = dict()
-    url_kwargs['callback'] = callback
-    if settings.GOOGLE_MAPS_API_KEY:
-        url_kwargs['key'] = settings.GOOGLE_MAPS_API_KEY
-
-    url = url + '?' + urlencode(url_kwargs)
-
-    return url
 
 
 # jsonify
