+++ project/map/templates/map/map_scripts.html
@@ -1,4 +1,5 @@
-{% load google_maps_api_url jsonify from common_tags %}
+{% load jsonify from common_tags %}
+{% load google_maps_api_url from map %}
 
 <script type="text/javascript">
   {# This function should be called after Google Maps loads. #}
